refactor(query): Extract query_parse/parser.rs into filters, ordering, aggregates, mutations
Extract filter parsing to filters.rs, order parsing to ordering.rs,
aggregate/groupBy parsing to aggregates.rs, mutation parsing to mutations.rs.

parser.rs: 3,240 → 2,240 lines (~1,100 lines of tests remain)
