+++ crates/query/src/query_parse/parser.rs
@@ -7,16 +7,17 @@ use graphql_parser::query::{
     SelectionSet, Value, VariableDefinition,
 };
 use serde_json::Value as JsonValue;
-use std::collections::{BTreeMap, HashMap, HashSet};
+use std::collections::{HashMap, HashSet};
 use tracing::instrument;
 
 use crate::document::DocumentMapping;
 use crate::error::{QueryError, Result};
-use crate::mapper::{
-    parse_mutation_name, Aggregate, AggregateTarget, AggregateType, Field as SelectField, Filter,
-    GroupBy, Limit, Mutation, MutationType, OrderBy, OrderCondition, OrderDirection, Requestable,
-    Select, Similarity,
-};
+use crate::mapper::{AggregateType, Field as SelectField, Limit, Mutation, Requestable, Select};
+
+use super::aggregates::{parse_aggregate_field, parse_group_by_value, parse_top_level_aggregate};
+use super::filters::parse_filter_value;
+use super::mutations::{parse_field_to_mutation, parse_similarity_field};
+use super::ordering::parse_order_value;
 
 /// Type of explain output requested.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
@@ -120,7 +121,7 @@ fn check_field_explain_directive(directives: &[Directive<'_, String>]) -> Result
 }
 
 /// Type alias for fragment definitions map
-type FragmentMap<'a> = HashMap<String, &'a FragmentDefinition<'a, String>>;
+pub(super) type FragmentMap<'a> = HashMap<String, &'a FragmentDefinition<'a, String>>;
 
 /// Parse a selection into Select operations, handling fragments.
 fn parse_selection_to_selects<'a>(
@@ -685,7 +686,7 @@ fn parse_field_to_select(
 }
 
 /// Parse a selection set into fields and document mapping.
-fn parse_selection_set(
+pub(super) fn parse_selection_set(
     selection_set: &SelectionSet<'_, String>,
     _collection_name: &str,
     variables: Option<&HashMap<String, JsonValue>>,
@@ -856,56 +857,6 @@ fn parse_selection_set(
     Ok((fields, mapping))
 }
 
-/// Parse a filter argument value into a Filter.
-fn parse_filter_value(
-    value: &Value<'_, String>,
-    variables: Option<&HashMap<String, JsonValue>>,
-) -> Result<Filter> {
-    match value {
-        Value::Object(obj) => {
-            let conditions = parse_filter_object(obj, variables)?;
-            Ok(Filter::from_conditions(conditions))
-        }
-        Value::Variable(name) => {
-            let vars = variables.ok_or_else(|| {
-                QueryError::parse(format!(
-                    "variable '{}' used but no variables provided",
-                    name
-                ))
-            })?;
-            let json_val = vars.get(name.as_str()).ok_or_else(|| {
-                QueryError::parse(format!("Variable \"${}\" was not provided", name))
-            })?;
-            if let JsonValue::Object(obj) = json_val {
-                let conditions: HashMap<String, JsonValue> =
-                    obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect();
-                Ok(Filter::from_conditions(conditions))
-            } else {
-                Err(QueryError::parse("filter must be an object"))
-            }
-        }
-        _ => Err(QueryError::parse("filter must be an object")),
-    }
-}
-
-/// Parse a filter object into conditions map.
-fn parse_filter_object(
-    obj: &BTreeMap<String, Value<'_, String>>,
-    variables: Option<&HashMap<String, JsonValue>>,
-) -> Result<HashMap<String, JsonValue>> {
-    let mut conditions = HashMap::new();
-
-    for (key, val) in obj {
-        let json_val = graphql_value_to_json(val, variables)?;
-        conditions.insert(key.clone(), json_val);
-    }
-
-    Ok(conditions)
-}
-
-/// Merge provided variables with default values.
-///
-/// Provided variables take precedence over defaults.
 fn merge_variables(
     provided: Option<&HashMap<String, JsonValue>>,
     defaults: &HashMap<String, JsonValue>,
@@ -973,7 +924,7 @@ fn graphql_value_to_json_no_vars(value: &Value<'_, String>) -> Result<JsonValue>
 }
 
 /// Convert GraphQL Value to JSON Value, resolving variables if present.
-fn graphql_value_to_json(
+pub(super) fn graphql_value_to_json(
     value: &Value<'_, String>,
     variables: Option<&HashMap<String, JsonValue>>,
 ) -> Result<JsonValue> {
@@ -1018,7 +969,7 @@ fn graphql_value_to_json(
 }
 
 /// Parse an integer value from GraphQL Value, resolving variables if present.
-fn parse_int_value(
+pub(super) fn parse_int_value(
     value: &Value<'_, String>,
     variables: Option<&HashMap<String, JsonValue>>,
 ) -> Result<i64> {
@@ -1078,547 +1029,8 @@ fn parse_optional_int_value(
     }
 }
 
-/// Parse order argument into OrderBy.
-/// Supports both single object `{field: ASC}` and array `[{field: ASC}, {other: DESC}]` formats.
-fn parse_order_value(
-    value: &Value<'_, String>,
-    variables: Option<&HashMap<String, JsonValue>>,
-) -> Result<OrderBy> {
-    let mut order_by = OrderBy::new();
-
-    match value {
-        Value::Object(obj) => {
-            // Go DefraDB requires each order argument to only define one field
-            if obj.len() > 1 {
-                return Err(QueryError::parse(
-                    "each order argument can only define one field",
-                ));
-            }
-            for (field_name, direction_val) in obj {
-                if let Some(condition) =
-                    parse_order_condition(field_name.clone(), direction_val, variables)?
-                {
-                    order_by = order_by.with_condition(condition);
-                }
-            }
-        }
-        Value::List(items) => {
-            // Array of order objects: [{rating: ASC}, {publisher: {yearOpened: DESC}}]
-            for item in items {
-                if let Value::Object(obj) = item {
-                    if obj.len() > 1 {
-                        return Err(QueryError::parse(
-                            "each order argument can only define one field",
-                        ));
-                    }
-                    for (field_name, direction_val) in obj {
-                        if let Some(condition) =
-                            parse_order_condition(field_name.clone(), direction_val, variables)?
-                        {
-                            order_by = order_by.with_condition(condition);
-                        }
-                    }
-                } else {
-                    return Err(QueryError::parse(
-                        "each order item in array must be an object",
-                    ));
-                }
-            }
-        }
-        Value::Variable(name) => {
-            let vars = variables.ok_or_else(|| {
-                QueryError::parse(format!(
-                    "variable '{}' used but no variables provided",
-                    name
-                ))
-            })?;
-            let json_val = vars.get(name).ok_or_else(|| {
-                QueryError::parse(format!("Variable \"${}\" was not provided", name))
-            })?;
-            return parse_order_from_json(json_val);
-        }
-        _ => return Err(QueryError::parse("order must be an object or array")),
-    }
-
-    Ok(order_by)
-}
-
-/// Parse order from a resolved JSON variable value.
-fn parse_order_from_json(json: &JsonValue) -> Result<OrderBy> {
-    let mut order_by = OrderBy::new();
-    match json {
-        JsonValue::Object(obj) => {
-            for (field_name, dir_val) in obj {
-                if let Some(dir_str) = dir_val.as_str() {
-                    if let Some(direction) = OrderDirection::parse(dir_str) {
-                        order_by =
-                            order_by.with_condition(OrderCondition::new(field_name, direction));
-                    }
-                }
-            }
-        }
-        JsonValue::Array(items) => {
-            for item in items {
-                if let JsonValue::Object(obj) = item {
-                    for (field_name, dir_val) in obj {
-                        if let Some(dir_str) = dir_val.as_str() {
-                            if let Some(direction) = OrderDirection::parse(dir_str) {
-                                order_by = order_by
-                                    .with_condition(OrderCondition::new(field_name, direction));
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        _ => {
-            return Err(QueryError::parse(
-                "order variable must be an object or array",
-            ))
-        }
-    }
-    Ok(order_by)
-}
-
-/// Parse a single order condition, handling nested relation ordering.
-/// Supports both simple `{field: ASC}` and nested `{relation: {field: DESC}}`.
-/// Returns None for null values (order field is ignored).
-fn parse_order_condition(
-    field_name: String,
-    direction_val: &Value<'_, String>,
-    variables: Option<&HashMap<String, JsonValue>>,
-) -> Result<Option<OrderCondition>> {
-    match direction_val {
-        // Null order direction means skip this field (Go compatibility)
-        Value::Null => Ok(None),
-        Value::Enum(s) | Value::String(s) => {
-            let direction = OrderDirection::parse(s).ok_or_else(|| {
-                // Match Go DefraDB error format
-                QueryError::parse(format!(
-                    "Argument \"order\" has invalid value {{{}: {}}}",
-                    field_name, s
-                ))
-            })?;
-            Ok(Some(OrderCondition::new(field_name, direction)))
-        }
-        Value::Variable(name) => {
-            let vars = variables.ok_or_else(|| {
-                QueryError::parse(format!(
-                    "variable '{}' used but no variables provided",
-                    name
-                ))
-            })?;
-            let json_val = vars.get(name).ok_or_else(|| {
-                QueryError::parse(format!("Variable \"${}\" was not provided", name))
-            })?;
-            let s = json_val.as_str().ok_or_else(|| {
-                QueryError::parse(format!(
-                    "Variable \"${}\" must be of type Ordering (ASC or DESC)",
-                    name
-                ))
-            })?;
-            let direction = OrderDirection::parse(s).ok_or_else(|| {
-                // Match Go DefraDB error format
-                QueryError::parse(format!(
-                    "Argument \"order\" has invalid value {{{}: {}}}",
-                    field_name, s
-                ))
-            })?;
-            Ok(Some(OrderCondition::new(field_name, direction)))
-        }
-        Value::Object(nested_obj) => {
-            // Nested ordering: {relation: {field: ASC}} or {_alias: {aliasName: ASC}}
-            // Empty nested order is a no-op (Go compatibility)
-            if nested_obj.is_empty() {
-                return Ok(None);
-            }
-            // Recursively parse the nested object
-            if nested_obj.len() != 1 {
-                return Err(QueryError::parse(
-                    "nested order must have exactly one field",
-                ));
-            }
-            let (nested_field, nested_direction) = nested_obj.iter().next().unwrap();
-            let nested_condition =
-                parse_order_condition(nested_field.clone(), nested_direction, variables)?;
-            // If nested condition is None (null value), propagate the None
-            match nested_condition {
-                Some(mut cond) => {
-                    // Handle _alias directive: don't prepend "_alias", just use the nested field name.
-                    // This allows ordering by aliased fields like: order: {_alias: {MyAge: ASC}}
-                    // where MyAge is an alias for the Age field.
-                    if field_name != "_alias" {
-                        // For regular nested ordering (relations), prepend the parent field to the path
-                        cond.fields.insert(0, field_name);
-                    }
-                    Ok(Some(cond))
-                }
-                None => Ok(None),
-            }
-        }
-        _ => Err(QueryError::parse("order direction must be ASC or DESC")),
-    }
-}
-
-/// Parse groupBy argument into GroupBy.
-fn parse_group_by_value(
-    value: &Value<'_, String>,
-    variables: Option<&HashMap<String, JsonValue>>,
-) -> Result<GroupBy> {
-    match value {
-        Value::List(items) => {
-            let fields: Result<Vec<String>> = items
-                .iter()
-                .map(|v| match v {
-                    Value::String(s) | Value::Enum(s) => Ok(s.clone()),
-                    Value::Variable(name) => {
-                        let vars = variables.ok_or_else(|| {
-                            QueryError::parse(format!(
-                                "variable '{}' used but no variables provided",
-                                name
-                            ))
-                        })?;
-                        let json_val = vars.get(name).ok_or_else(|| {
-                            QueryError::parse(format!("Variable \"${}\" was not provided", name))
-                        })?;
-                        json_val.as_str().map(|s| s.to_string()).ok_or_else(|| {
-                            QueryError::parse(format!(
-                                "Variable \"${}\" must be of type String",
-                                name
-                            ))
-                        })
-                    }
-                    _ => Err(QueryError::parse("groupBy items must be strings")),
-                })
-                .collect();
-            Ok(GroupBy::new(fields?))
-        }
-        Value::Variable(name) => {
-            let vars = variables.ok_or_else(|| {
-                QueryError::parse(format!(
-                    "variable '{}' used but no variables provided",
-                    name
-                ))
-            })?;
-            let json_val = vars.get(name).ok_or_else(|| {
-                QueryError::parse(format!("Variable \"${}\" was not provided", name))
-            })?;
-            let arr = json_val.as_array().ok_or_else(|| {
-                QueryError::parse(format!("Variable \"${}\" must be of type [String]", name))
-            })?;
-            let fields: Result<Vec<String>> = arr
-                .iter()
-                .map(|v| {
-                    v.as_str()
-                        .map(|s| s.to_string())
-                        .ok_or_else(|| QueryError::parse("groupBy items must be strings"))
-                })
-                .collect();
-            Ok(GroupBy::new(fields?))
-        }
-        _ => Err(QueryError::parse("groupBy must be a list")),
-    }
-}
-
-/// Parse an aggregate target from a GraphQL Object value.
-fn parse_aggregate_target_obj(
-    arg_name: &str,
-    obj: &BTreeMap<String, Value<'_, String>>,
-    variables: Option<&HashMap<String, JsonValue>>,
-) -> Result<AggregateTarget> {
-    let mut target = AggregateTarget::new(arg_name.to_string());
-    for (key, val) in obj {
-        match key.as_str() {
-            "field" => {
-                let field_name = match val {
-                    Value::String(s) => s.clone(),
-                    Value::Enum(s) => s.clone(),
-                    _ => {
-                        return Err(QueryError::parse(
-                            "field in relation aggregate must be a string",
-                        ))
-                    }
-                };
-                target.field_name = Some(field_name);
-            }
-            "filter" => {
-                let filter = parse_filter_value(val, variables)?;
-                target.filter = Some(filter);
-            }
-            "limit" => {
-                let limit_val = parse_int_value(val, variables)?;
-                target.limit = Some(Limit::new(Some(limit_val as u64), 0));
-            }
-            "offset" => {
-                let offset_val = parse_int_value(val, variables)?;
-                if let Some(ref mut limit) = target.limit {
-                    limit.offset = offset_val as u64;
-                } else {
-                    target.limit = Some(Limit::new(None, offset_val as u64));
-                }
-            }
-            "order" => {
-                let order = match val {
-                    Value::Enum(s) | Value::String(s) => {
-                        let direction = OrderDirection::parse(s).ok_or_else(|| {
-                            QueryError::parse(format!("invalid order direction: {}", s))
-                        })?;
-                        OrderBy::new().with_condition(OrderCondition::new("", direction))
-                    }
-                    _ => parse_order_value(val, variables)?,
-                };
-                target.order = Some(order);
-            }
-            _ => {}
-        }
-    }
-    Ok(target)
-}
-
-/// Parse an aggregate target from a resolved JSON variable value.
-fn parse_aggregate_target_from_json(
-    arg_name: &str,
-    json: &JsonValue,
-    _variables: Option<&HashMap<String, JsonValue>>,
-) -> Result<AggregateTarget> {
-    let mut target = AggregateTarget::new(arg_name.to_string());
-    if let JsonValue::Object(obj) = json {
-        for (key, val) in obj {
-            match key.as_str() {
-                "field" => {
-                    if let Some(s) = val.as_str() {
-                        target.field_name = Some(s.to_string());
-                    }
-                }
-                "filter" => {
-                    // Convert JSON filter to a Filter
-                    if let JsonValue::Object(filter_obj) = val {
-                        let conditions: HashMap<String, JsonValue> = filter_obj
-                            .iter()
-                            .map(|(k, v)| (k.clone(), v.clone()))
-                            .collect();
-                        target.filter = Some(Filter::from_conditions(conditions));
-                    }
-                }
-                "limit" => {
-                    if let Some(n) = val.as_i64() {
-                        target.limit = Some(Limit::new(Some(n as u64), 0));
-                    }
-                }
-                "offset" => {
-                    if let Some(n) = val.as_i64() {
-                        if let Some(ref mut limit) = target.limit {
-                            limit.offset = n as u64;
-                        } else {
-                            target.limit = Some(Limit::new(None, n as u64));
-                        }
-                    }
-                }
-                "order" => {
-                    target.order = Some(parse_order_from_json(val)?);
-                }
-                _ => {}
-            }
-        }
-    }
-    Ok(target)
-}
-
-/// Parse an aggregate field into an Aggregate.
-///
-/// Handles aggregate functions like `_count`, `_sum(field: "age")`, etc.
-/// Also supports relation aggregates like `_count(books: {})`, `_sum(books: {field: score}, articles: {field: rating})`.
-fn parse_aggregate_field(
-    field: &Field<'_, String>,
-    agg_type: AggregateType,
-    variables: Option<&HashMap<String, JsonValue>>,
-) -> Result<Aggregate> {
-    let mut target_field: Option<String> = None;
-    let mut relation_targets: Vec<AggregateTarget> = Vec::new();
-
-    // Parse arguments (e.g., `field: "age"` for _sum, or `books: {}` for relation aggregates)
-    for (arg_name, arg_value) in &field.arguments {
-        match arg_name.as_str() {
-            "field" => {
-                target_field = Some(match arg_value {
-                    Value::String(s) => s.clone(),
-                    Value::Enum(s) => s.clone(),
-                    Value::Variable(name) => {
-                        let vars = variables.ok_or_else(|| {
-                            QueryError::parse(format!(
-                                "variable '{}' used but no variables provided",
-                                name
-                            ))
-                        })?;
-                        let json_val = vars.get(name).ok_or_else(|| {
-                            QueryError::parse(format!("Variable \"${}\" was not provided", name))
-                        })?;
-                        json_val
-                            .as_str()
-                            .ok_or_else(|| {
-                                QueryError::parse(format!(
-                                    "Variable \"${}\" must be of type String",
-                                    name
-                                ))
-                            })?
-                            .to_string()
-                    }
-                    _ => return Err(QueryError::parse("field argument must be a string")),
-                });
-            }
-            _ => {
-                // This might be a relation name argument like `books: {field: score}`
-                // Relation arguments take an object value with optional field, filter, limit, order.
-                // Also handle variables that resolve to objects.
-                match arg_value {
-                    Value::Object(obj) => {
-                        let target = parse_aggregate_target_obj(arg_name, obj, variables)?;
-                        relation_targets.push(target);
-                    }
-                    Value::Variable(name) => {
-                        let vars = variables.ok_or_else(|| {
-                            QueryError::parse(format!(
-                                "variable '{}' used but no variables provided",
-                                name
-                            ))
-                        })?;
-                        let json_val = vars.get(name).ok_or_else(|| {
-                            QueryError::parse(format!("Variable \"${}\" was not provided", name))
-                        })?;
-                        let target =
-                            parse_aggregate_target_from_json(arg_name, json_val, variables)?;
-                        relation_targets.push(target);
-                    }
-                    _ => {
-                        return Err(QueryError::parse(format!(
-                            "unknown argument '{}' on aggregate '{}', expected an object value for relation aggregates",
-                            arg_name,
-                            agg_type.as_str()
-                        )));
-                    }
-                }
-            }
-        }
-    }
-
-    // Create the appropriate aggregate
-    let aggregate = if !relation_targets.is_empty() {
-        // Relation aggregate mode - create aggregate with targets directly (no empty initial target)
-        Aggregate {
-            aggregate_type: agg_type,
-            targets: relation_targets,
-            filter: None,
-            alias: None,
-        }
-    } else {
-        // Simple field aggregate mode
-        match agg_type {
-            AggregateType::Count => {
-                // _count can work without a field argument (counts all docs)
-                if let Some(field_name) = target_field {
-                    Aggregate::count().with_target(AggregateTarget::with_field("", field_name))
-                } else {
-                    Aggregate::count()
-                }
-            }
-            AggregateType::Sum => {
-                let field_name = target_field.ok_or_else(|| {
-                    QueryError::parse("_sum requires a 'field' argument or relation targets")
-                })?;
-                Aggregate::sum(AggregateTarget::with_field("", field_name))
-            }
-            AggregateType::Average => {
-                let field_name = target_field.ok_or_else(|| {
-                    QueryError::parse("_avg requires a 'field' argument or relation targets")
-                })?;
-                Aggregate::avg(AggregateTarget::with_field("", field_name))
-            }
-            AggregateType::Min => {
-                let field_name = target_field.ok_or_else(|| {
-                    QueryError::parse("_min requires a 'field' argument or relation targets")
-                })?;
-                Aggregate::min(AggregateTarget::with_field("", field_name))
-            }
-            AggregateType::Max => {
-                let field_name = target_field.ok_or_else(|| {
-                    QueryError::parse("_max requires a 'field' argument or relation targets")
-                })?;
-                Aggregate::max(AggregateTarget::with_field("", field_name))
-            }
-        }
-    };
-
-    Ok(aggregate)
-}
-
-/// Parse a top-level aggregate query (e.g., `{ _avg(Users: {field: Age}) }`).
-///
-/// Top-level aggregates are different from nested aggregates in that:
-/// - The aggregate function name is the top-level field
-/// - Arguments are collection names with their aggregate configuration
-/// - The result is wrapped in a Select with the collection as the target
-fn parse_top_level_aggregate(
-    field: &Field<'_, String>,
-    agg_type: AggregateType,
-    variables: Option<&HashMap<String, JsonValue>>,
-) -> Result<Select> {
-    let mut aggregate = parse_aggregate_field(field, agg_type, variables)?;
-
-    // Top-level numeric aggregates require a field argument on each target.
-    // This matches Go's GraphQL schema validation where collection args require
-    // the "field" key (e.g., _avg(Users: {field: Age}) not _avg(Users: {})).
-    if agg_type != AggregateType::Count {
-        for target in &aggregate.targets {
-            if target.field_name.is_none() {
-                let type_name = format!("{}NumericFieldsArg!", capitalize_first(&target.host_name));
-                return Err(QueryError::parse(format!(
-                    "Argument \"{}\" has invalid value {{}}.\nIn field \"field\": Expected \"{}\", found null.",
-                    target.host_name,
-                    type_name
-                )));
-            }
-        }
-    }
-
-    // Set alias if provided
-    if let Some(ref a) = field.alias {
-        aggregate = aggregate.with_alias(a.clone());
-    }
-
-    // Get the collection name from the first target
-    let collection_name = aggregate
-        .targets
-        .first()
-        .map(|t| t.host_name.clone())
-        .unwrap_or_else(|| String::new());
-
-    // Create a Select that wraps this aggregate
-    // The field name should be the aggregate name (e.g., "_avg") so the response
-    // key is correct (e.g., {"_avg": 29} not {"Users": 29})
-    let mut select = Select::new(&collection_name);
-    let agg_name = agg_type.as_str();
-    if let Some(ref a) = field.alias {
-        // If aliased, use alias as the output name: { average: _avg(...) } -> {"average": ...}
-        select.field = SelectField::with_alias(agg_name, a.clone());
-    } else {
-        // Otherwise use the aggregate name: { _avg(...) } -> {"_avg": ...}
-        select.field = SelectField::new(agg_name);
-    }
-
-    // Add to document mapping
-    let index = select.document_mapping.next_index();
-    select.document_mapping.add(index, agg_name);
-    select
-        .document_mapping
-        .add_render_key(index, aggregate.output_name());
-
-    select.fields.push(Requestable::Aggregate(aggregate));
-
-    Ok(select)
-}
-
 /// Parse docIDs argument into vector of strings.
-fn parse_doc_ids_value(
+pub(super) fn parse_doc_ids_value(
     value: &Value<'_, String>,
     variables: Option<&HashMap<String, JsonValue>>,
 ) -> Result<Vec<String>> {
@@ -1741,422 +1153,10 @@ fn resolve_bool_value(
     }
 }
 
-// =============================================================================
-// Mutation Parsing
-// =============================================================================
-
-/// Parse a single GraphQL field into a Mutation operation.
-///
-/// Mutation field names follow the format: `operation_collection`
-/// Examples: `create_Users`, `update_Posts`, `delete_Comments`
-fn parse_field_to_mutation(
-    field: &Field<'_, String>,
-    variables: Option<&HashMap<String, JsonValue>>,
-) -> Result<Mutation> {
-    let field_name = &field.name;
-
-    // Parse mutation name to get operation type and collection
-    let (mutation_type, collection_name) =
-        parse_mutation_name(field_name).map_err(QueryError::parse)?;
-
-    // Create base mutation
-    let mut mutation = match mutation_type {
-        MutationType::Create => Mutation::create(&collection_name),
-        MutationType::Update => Mutation::update(&collection_name),
-        MutationType::Delete => Mutation::delete(&collection_name),
-        MutationType::Upsert => Mutation::upsert(&collection_name),
-    };
-
-    // Capture GraphQL alias if present (e.g., "john: update_Users(...)")
-    if let Some(ref alias) = field.alias {
-        mutation = mutation.with_alias(alias.clone());
-    }
-
-    // Track if input argument was present (even if null)
-    let mut has_input_arg = false;
-
-    // Parse arguments based on mutation type
-    for (arg_name, arg_value) in &field.arguments {
-        match (mutation_type, arg_name.as_str()) {
-            // CREATE: input is array of documents (null means empty operation)
-            (MutationType::Create, "input") => {
-                has_input_arg = true;
-                if !matches!(arg_value, Value::Null) {
-                    let input = parse_create_input(arg_value, variables)?;
-                    mutation.create_input = input;
-                }
-                // null input is valid - leaves create_input empty for empty result
-            }
-
-            // UPDATE: input is patch object (null means empty operation)
-            (MutationType::Update, "input") => {
-                has_input_arg = true;
-                if !matches!(arg_value, Value::Null) {
-                    let input = parse_update_input(arg_value, variables)?;
-                    mutation.update_input = input;
-                }
-            }
-
-            // UPSERT: create is the document to create if no match (single object, not array)
-            (MutationType::Upsert, "create") => {
-                if matches!(arg_value, Value::Null) {
-                    return Err(QueryError::parse(
-                        "Argument \"create\" has invalid value <nil>.".to_string(),
-                    ));
-                }
-                let input = parse_update_input(arg_value, variables)?;
-                // Store create input as single-element array for consistency
-                mutation.create_input = vec![input];
-            }
-
-            // UPSERT: update is the fields to update if match found
-            (MutationType::Upsert, "update") => {
-                if matches!(arg_value, Value::Null) {
-                    return Err(QueryError::parse(
-                        "Argument \"update\" has invalid value <nil>.".to_string(),
-                    ));
-                }
-                let input = parse_update_input(arg_value, variables)?;
-                mutation.update_input = input;
-            }
-
-            // UPDATE/DELETE: docID or docIDs to target (Go uses singular docID)
-            (MutationType::Update | MutationType::Delete, "docID" | "docIDs" | "_docIDs") => {
-                // Null docIDs is valid and means "no specific docIDs" (use filter or all)
-                if !matches!(arg_value, Value::Null) {
-                    let doc_ids = parse_doc_ids_value(arg_value, variables)?;
-                    mutation.doc_ids = Some(doc_ids);
-                }
-            }
-
-            // UPDATE/DELETE/UPSERT: filter to find documents
-            (MutationType::Update | MutationType::Delete | MutationType::Upsert, "filter") => {
-                // Null filter is valid and means "no filter" (operate on all docs)
-                if !matches!(arg_value, Value::Null) {
-                    let filter = parse_filter_value(arg_value, variables)?;
-                    mutation.filter = Some(filter);
-                }
-            }
-
-            // Encryption: encrypt entire document
-            (_, "encrypt") => {
-                if let Value::Boolean(b) = arg_value {
-                    mutation.encrypt_doc = *b;
-                }
-            }
-
-            // Encryption: encrypt specific fields
-            (_, "encryptFields") => {
-                if let Value::List(fields) = arg_value {
-                    mutation.encrypt_fields = fields
-                        .iter()
-                        .filter_map(|v| match v {
-                            Value::Enum(name) => Some(name.clone()),
-                            Value::String(name) => Some(name.clone()),
-                            _ => None,
-                        })
-                        .collect();
-                }
-            }
-
-            // Unknown argument
-            _ => {
-                return Err(QueryError::parse(format!(
-                    "Unknown argument '{}' for {} mutation on '{}'",
-                    arg_name,
-                    mutation_type.as_prefix(),
-                    collection_name
-                )));
-            }
-        }
-    }
-
-    // Validate mutation has required arguments
-    match mutation_type {
-        MutationType::Create => {
-            if mutation.create_input.is_empty() && !has_input_arg {
-                return Err(QueryError::parse(format!(
-                    "create_{} mutation requires 'input' argument with array of documents",
-                    collection_name
-                )));
-            }
-        }
-        MutationType::Update => {
-            if mutation.update_input.is_empty() && !has_input_arg {
-                return Err(QueryError::parse(format!(
-                    "update_{} mutation requires 'input' argument with fields to update",
-                    collection_name
-                )));
-            }
-            // Note: Go DefraDB allows update without docIDs or filter
-            // (meaning update all documents in the collection)
-        }
-        MutationType::Delete => {
-            // Note: Go DefraDB allows delete without docIDs or filter
-            // (meaning delete all documents in the collection)
-        }
-        MutationType::Upsert => {
-            // Go DefraDB requires all three: filter, create, update
-            if mutation.filter.is_none() {
-                return Err(QueryError::parse(format!(
-                    "upsert_{} mutation requires 'filter' argument",
-                    collection_name
-                )));
-            }
-            if mutation.create_input.is_empty() {
-                return Err(QueryError::parse(format!(
-                    "upsert_{} mutation requires 'create' argument with document to create if no match",
-                    collection_name
-                )));
-            }
-            if mutation.update_input.is_empty() {
-                return Err(QueryError::parse(format!(
-                    "upsert_{} mutation requires 'update' argument with fields to update if match found",
-                    collection_name
-                )));
-            }
-        }
-    }
-
-    // Parse selection set (fields to return after mutation)
-    // For mutations, we don't support fragments in return fields
-    let empty_fragments: FragmentMap<'_> = HashMap::new();
-    let mut empty_visiting = HashSet::new();
-    let (fields, mapping) = parse_selection_set(
-        &field.selection_set,
-        &collection_name,
-        variables,
-        &empty_fragments,
-        &mut empty_visiting,
-    )?;
-
-    // All mutations return [TypeName], which is an object type requiring a sub selection.
-    if fields.is_empty() {
-        return Err(QueryError::parse(format!(
-            "Field \"{}\" of type \"[{}]\" must have a sub selection.",
-            field_name, collection_name
-        )));
-    }
-
-    mutation.fields = fields;
-    mutation.document_mapping = mapping;
-
-    Ok(mutation)
-}
-
-/// Parse CREATE mutation input (array of documents).
-fn parse_create_input(
-    value: &Value<'_, String>,
-    variables: Option<&HashMap<String, JsonValue>>,
-) -> Result<Vec<HashMap<String, JsonValue>>> {
-    match value {
-        Value::List(items) => {
-            let mut docs = Vec::new();
-            for item in items {
-                match item {
-                    Value::Object(obj) => {
-                        let doc = parse_document_input(obj, variables)?;
-                        docs.push(doc);
-                    }
-                    _ => return Err(QueryError::parse("CREATE input items must be objects")),
-                }
-            }
-            Ok(docs)
-        }
-        Value::Object(obj) => {
-            // Single document (wrap in array)
-            let doc = parse_document_input(obj, variables)?;
-            Ok(vec![doc])
-        }
-        // Null input is valid and means "no documents to create"
-        Value::Null => Ok(vec![]),
-        // Variable reference - resolve from variables map
-        Value::Variable(var_name) => {
-            let vars = variables.ok_or_else(|| {
-                QueryError::parse(format!(
-                    "variable '{}' used but no variables provided",
-                    var_name
-                ))
-            })?;
-            let json_val = vars.get(var_name).ok_or_else(|| {
-                QueryError::parse(format!("variable '{}' not found in variables", var_name))
-            })?;
-            // Convert JSON value to documents
-            match json_val {
-                JsonValue::Array(items) => {
-                    let mut docs = Vec::new();
-                    for item in items {
-                        if let JsonValue::Object(obj) = item {
-                            let doc: HashMap<String, JsonValue> =
-                                obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect();
-                            docs.push(doc);
-                        } else {
-                            return Err(QueryError::parse("CREATE input items must be objects"));
-                        }
-                    }
-                    Ok(docs)
-                }
-                JsonValue::Object(obj) => {
-                    let doc: HashMap<String, JsonValue> =
-                        obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect();
-                    Ok(vec![doc])
-                }
-                JsonValue::Null => Ok(vec![]),
-                _ => Err(QueryError::parse(
-                    "CREATE input variable must be an array of objects or a single object",
-                )),
-            }
-        }
-        _ => Err(QueryError::parse(
-            "CREATE input must be an array of objects or a single object",
-        )),
-    }
-}
-
-/// Parse UPDATE mutation input (patch object).
-/// Non-object input (e.g., array "patch") is treated as empty/no-op (Go compatibility).
-fn parse_update_input(
-    value: &Value<'_, String>,
-    variables: Option<&HashMap<String, JsonValue>>,
-) -> Result<HashMap<String, JsonValue>> {
-    match value {
-        Value::Object(obj) => parse_document_input(obj, variables),
-        _ => Ok(HashMap::new()),
-    }
-}
-
-/// Parse a document input object into field-value map.
-fn parse_document_input(
-    obj: &BTreeMap<String, Value<'_, String>>,
-    variables: Option<&HashMap<String, JsonValue>>,
-) -> Result<HashMap<String, JsonValue>> {
-    let mut fields = HashMap::new();
-    for (key, value) in obj {
-        let json_value = graphql_value_to_json(value, variables)?;
-        fields.insert(key.clone(), json_value);
-    }
-    Ok(fields)
-}
-
-/// Capitalize the first character of a string.
-fn capitalize_first(s: &str) -> String {
-    let mut chars = s.chars();
-    match chars.next() {
-        Some(c) => c.to_uppercase().to_string() + chars.as_str(),
-        None => String::new(),
-    }
-}
-
-/// Parse a _similarity field from a GraphQL query.
-///
-/// Format: `_similarity(fieldName: {vector: [1, 2, 3]})`
-/// The argument name is the target field containing the document's vector.
-/// The value is an object with a `vector` key containing the query vector.
-fn parse_similarity_field(
-    field: &Field<'_, String>,
-    variables: Option<&HashMap<String, JsonValue>>,
-) -> Result<Similarity> {
-    if field.arguments.is_empty() {
-        return Err(QueryError::parse("_similarity requires a field argument"));
-    }
-
-    let (target_field, value) = &field.arguments[0];
-
-    // Parse the value: {vector: [1, 2, 3]}
-    let vector = match value {
-        Value::Object(obj) => {
-            let vec_value = obj.get("vector").ok_or_else(|| {
-                QueryError::parse("_similarity argument must contain a 'vector' key")
-            })?;
-            parse_vector_value(vec_value, variables)?
-        }
-        Value::Variable(var_name) => {
-            let vars = variables.ok_or_else(|| {
-                QueryError::parse(format!(
-                    "variable '{}' used but no variables provided",
-                    var_name
-                ))
-            })?;
-            let json_val = vars.get(var_name.as_str()).ok_or_else(|| {
-                QueryError::parse(format!("Variable \"${}\" was not provided", var_name))
-            })?;
-            if let JsonValue::Object(obj) = json_val {
-                let vec_val = obj.get("vector").ok_or_else(|| {
-                    QueryError::parse("_similarity variable must contain a 'vector' key")
-                })?;
-                parse_json_vector(vec_val)?
-            } else {
-                return Err(QueryError::parse("_similarity variable must be an object"));
-            }
-        }
-        _ => {
-            return Err(QueryError::parse(
-                "_similarity argument must be an object with 'vector' key",
-            ));
-        }
-    };
-
-    Ok(Similarity::new(target_field.clone(), vector))
-}
-
-/// Parse a vector value from a GraphQL list literal.
-fn parse_vector_value(
-    value: &Value<'_, String>,
-    _variables: Option<&HashMap<String, JsonValue>>,
-) -> Result<Vec<f64>> {
-    match value {
-        Value::List(items) => {
-            let mut vec = Vec::with_capacity(items.len());
-            for item in items {
-                match item {
-                    Value::Int(n) => {
-                        vec.push(
-                            n.as_i64()
-                                .ok_or_else(|| QueryError::parse("integer out of range"))?
-                                as f64,
-                        );
-                    }
-                    Value::Float(f) => {
-                        vec.push(*f);
-                    }
-                    _ => {
-                        return Err(QueryError::parse("vector values must be numeric"));
-                    }
-                }
-            }
-            Ok(vec)
-        }
-        _ => Err(QueryError::parse("vector must be an array")),
-    }
-}
-
-/// Parse a vector from a JSON array value.
-fn parse_json_vector(value: &JsonValue) -> Result<Vec<f64>> {
-    match value {
-        JsonValue::Array(items) => {
-            let mut vec = Vec::with_capacity(items.len());
-            for item in items {
-                match item {
-                    JsonValue::Number(n) => {
-                        vec.push(
-                            n.as_f64()
-                                .ok_or_else(|| QueryError::parse("invalid number in vector"))?,
-                        );
-                    }
-                    _ => {
-                        return Err(QueryError::parse("vector values must be numeric"));
-                    }
-                }
-            }
-            Ok(vec)
-        }
-        _ => Err(QueryError::parse("vector must be an array")),
-    }
-}
-
 #[cfg(test)]
 mod mutation_tests {
     use super::*;
+    use crate::mapper::MutationType;
 
     #[test]
     fn test_parse_create_mutation() {
